Move validation methods and props into a new HasValidation trait
